refactor: resolve circular imports with beartype integration
Extracts CodeQualityScores to a separate types module to prevent
circular import issues when runtime type checking is enabled.

Adds beartype configuration to package initialization with
environment variable control for global type checking behavior.

Removes TYPE_CHECKING conditional imports in favor of direct
imports from the new types module.
